update lua compile

+++ linux/graphictools/compile_rusticl/12_install_lua_5.4_latest.sh
@@ -42,6 +42,8 @@ install_lua_from_tarball() {
   local url="${LUA_BASE_URL}/${tarball}"
 
   log "📥 Downloading $tarball from $url"
+  [[ -z "$tarball" ]] && fail "Tarball name is empty"
+  [[ -z "$url" ]] && fail "URL is empty"
   rm -rf "$TMP_DIR"
   mkdir -p "$TMP_DIR"
   cd "$TMP_DIR"
